backend/requirements.txt
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0,<1.0
sqlalchemy>=2.0.0,<3.0
pydantic>=2.0.0,<3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
requests>=2.31.0,<3.0
aiohttp>=3.9.0,<4.0
paramiko>=3.4.0,<4.0
asyncssh>=2.14.0,<3.0
psutil>=5.9.0,<6.0
prometheus-client>=0.19.0,<1.0
python-jose>=3.3.0,<4.0
passlib>=1.7.4,<2.0
bcrypt>=4.0.0,<5.0
croniter>=2.0.0,<3.0
# psycopg3 (psycopg[binary]) 有预编译 wheel，无需编译
psycopg[binary]>=3.1.0
hypothesis>=6.0.0,<7.0
pytest>=7.0.0,<9.0
pytest-asyncio>=0.23.0,<1.0
